alicevision-service/requirements.txt
# AliceVision Advanced AI Dependencies
# Professional-grade computer vision for AIWardrobe

# Core API
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.4.0

# Computer Vision & AI
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0

# Background removal & segmentation
rembg>=2.0.50

# Pose detection
mediapipe>=0.10.0

# NEW: Advanced AI Models - Option A (Grounded SAM2 + FashionCLIP)
# Grounded SAM2 - Text-prompted segmentation
git+https://github.com/IDEA-Research/GroundingDINO.git
git+https://github.com/facebookresearch/segment-anything-2.git
supervision>=0.16.0             # For visualization and tracking

# FashionCLIP - Specialized fashion understanding  
open-clip-torch>=2.24.0         # Fashion-CLIP for classification
sentence-transformers>=2.2.0    # Text embeddings

# Additional models
ultralytics>=8.0.0              # YOLOv8 as backup detector
faiss-cpu>=1.7.4                # Vector similarity search
scikit-learn>=1.3.0             # K-means clustering for colors
scikit-image>=0.21.0            # Image quality metrics
timm>=0.9.0                     # PyTorch image models

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0
